requirements.txt
PyQt6>=6.6.0
faster-whisper>=1.0.0
azure-cognitiveservices-speech>=1.35.0
vosk>=0.3.45
SpeechRecognition>=3.10.0
python-osc>=1.8.0
sounddevice>=0.4.6
PyAudio>=0.2.14
webrtcvad>=2.0.10
numpy>=1.24.0
platformdirs>=4.0.0
huggingface_hub>=0.20.0
soundfile>=0.12.0
pyttsx3>=2.90
keyboard>=0.13.5
boto3>=1.34.0
openvr>=1.26.701
